requirements.txt
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
distro==1.9.0
docstring_parser==0.17.0
filelock==3.19.1
filetype==1.2.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.4
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.21
langchain-core==0.3.78
langchain-google-genai==2.0.10
langchain-huggingface==0.3.1
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.32
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.3.3
openai==2.1.0
orjson==3.11.3
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==70.2.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
zstandard==0.25.0
ipykernel
streamlit
grandalf
PyPDF2
chromadb
faiss-cpu
langchain-ollama
langchain-community
watchdog
matplotlib
seaborn
plotly
rembg
cartooner
pymupdf
langchain-chroma
pypdf
wikipedia
ddgs